Workaround for odd stdout race
Data races never happen in Rust? Lollerz

If I don't put output in its own function,
the write seems to happen before the push completes.

That is, we don't get the ending "]" when we use the -e flag,
which is just weird, but eh.

The write should be its own function anyway, since I am mulling over
adding a -o flag to specify an output file. Sometimes using '>' is
a pain in the ass...